lrr_serv/lrr_serv/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-allauth==0.34.0
django-cors-headers==2.1.0
django-environ==0.4.4
django-rest-auth==0.9.2
djangorestframework==3.7.3
idna==2.6
oauthlib==2.0.6
python-openid==2.2.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
